stage2: last changes

+++ app/db/models.py
@@ -1,9 +1,5 @@
-from httpx._transports import default
-from sqlalchemy import Integer, Boolean, DateTime, Float
-from _pytest import compat
 from datetime import datetime
-
-from sqlalchemy import String, DateTime
+from sqlalchemy import Boolean, DateTime, Float, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.database import Base
@@ -11,15 +7,14 @@ from app.db.database import Base
 
 class Monitor(Base):
     __tablename__ = "monitors"
-    
+
     id: Mapped[int] = mapped_column(primary_key=True)
     name: Mapped[str] = mapped_column(String(255))
     url: Mapped[str] = mapped_column(String(255))
-    timeout: Mapped[int] = mapped_column(Float)
+    timeout: Mapped[float] = mapped_column(Float)
     interval: Mapped[int] = mapped_column(Integer)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     created_at: Mapped[datetime] = mapped_column(
         DateTime,
         default=datetime.utcnow
     )
-
